requirements.txt
absl-py==0.14.1
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
async-generator==1.10
attrs==21.2.0
autopep8==1.5.4
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
boto3==1.14.31
botocore==1.17.31
cachelib==0.1
certifi==2020.4.5.1
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
click==7.1.2
colorama==0.4.4
configparser==5.1.0
crayons==0.4.0
cryptography==36.0.1
cs50==5.0.4
cycler==0.10.0
distlib==0.3.1
Django==3.1.4
django-ckeditor==5.9.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.2.2
django-filter==2.3.0
django-js-asset==1.2.2
django-storages==1.9.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
filelock==3.0.12
Flask==1.1.2
Flask-Session==0.3.2
gunicorn==20.0.4
h11==0.12.0
idna==2.9
install==1.3.3
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kiwisolver==1.3.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.4.3
mccabe==0.6.1
mediapipe==0.8.8
numpy==1.21.2
opencv-contrib-python==4.5.3.56
outcome==1.1.0
Pillow==8.0.1
pipenv==2021.11.9
platformdirs==2.4.0
protobuf==3.18.1
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.21
PyJWT==1.7.1
pylint==2.6.0
pyodbc==4.0.32
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
python-nmap==0.7.1
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
selenium==4.0.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.3.17
sqlparse==0.3.1
termcolor==1.1.0
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
twilio==7.3.1
typing-extensions==3.10.0.2
urllib3==1.25.9
virtualenv==20.2.1
virtualenv-clone==0.5.7
webdriver-manager==3.4.0
Werkzeug==0.15.4
whitenoise==5.1.0
wrapt==1.12.1
wsproto==1.0.0
